+++ workbench/libs/codesetslib/developer/examples/b64e.c
@@ -56,7 +56,7 @@ int main(int argc,char **argv)
             r = CodesetsEncodeB64(CSA_B64SourceFile, (Tag)argv[1],
                                   CSA_B64DestFile,   (Tag)argv[2],
                                   TAG_DONE);
-            printf("Res %d\n",r);
+            printf("Res %d\n", (int)r);
 
             DROPINTERFACE(ICodesets);
             CloseLibrary(CodesetsBase);
